Fix PulsarConfigurationTests due to previous commit consolidation.
